add level filter logger to another file
